Create Reverse_String
This is a problem with reversing a string. First, we take two pointers, i and j, and initialize i with 0 and j with an n-1 index. 
Swap characters at i and jth index, incrementing i by 1 and decrementing j by 1. Until i greater than j (or) j less than i.
Finally, the string will be reversed.