Create Missing_number
The problem is to find the missing number. We solve this problem by taking the sum of n natural numbers.
First, take the sum of n natural numbers and then subtract the obtained sum from the sum of the elements of the given array. 
Finally, we get the missing number.